requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-allauth
psycopg2
gunicorn==19.6.0
twilio
pytz
python-decouple
pillow